[frontend] component to tsx (#12601)
